Added lib/logger as a submodule to lib/mojito
