+++ lib/mojito.js
@@ -11,6 +11,20 @@
 Main module in the Mojito package that is responsible for creating the mojito
 instance to the attached to the `express` app.
 
+Usage:
+
+    var app = require('express'),
+        mojito = require('mojito'),
+        app;
+
+    app = express();
+
+    app.use(require('./middleware/my-middleware'));
+
+    app.use(mojito.middleware(app));
+
+    app.listen(app.get('port'));
+
 @module mojito
 **/
 
@@ -22,9 +36,9 @@ var debug = require('debug')('mojito:server'),
     libpath = require('path'),
     appProto = express.application,
     defaultConfiguration = appProto.defaultConfiguration,
-    // OutputHandler = require('./output-handler.server'),
-    libstore = require('./store'),
+    liblogger = require('./logger'),
     libmiddleware = require('./middleware'),
+    libstore = require('./store'),
     Mojito;
 
 
@@ -41,6 +55,9 @@ function merge(a, b) {
         }
     }
 }
+/**
+Copied from modown-yui.utils
+**/
 function extend(obj) {
     Array.prototype.slice.call(arguments, 1).forEach(function (source) {
         var key;
@@ -55,7 +72,7 @@ function extend(obj) {
     });
 
     return obj;
-};
+}
 
 
 //  ----------------------------------------------------------------------------
@@ -89,26 +106,6 @@ Mojito.NOOP = function () {};
 
 
 /**
-An ordered list of the middleware module names to load for a standard Mojito
-server instance.
-@type Array
-**/
-/*
-Mojito.MOJITO_MIDDLEWARE = [
-    'mojito-handler-static',
-    'mojito-parser-body',
-    'mojito-parser-cookies',
-    'mojito-contextualizer',
-    'mojito-handler-tunnel',
-    'mojito-router',
-    'mojito-handler-dispatcher',
-    'mojito-handler-error'
-];
-*/
-
-
-
-/**
 Mojito extension for Express.
 
 @class ExpressMojitoExtension
@@ -124,8 +121,11 @@ function ExpressMojitoExtension(app) {
     return this;
 }
 
-// Mixin here
-ExpressMojitoExtension = extend(ExpressMojitoExtension, libmiddleware);
+// Mix in submodules here
+//
+// NOTE: This is unusual to mix in those submodules here, but the function 
+//       prototypes need them.
+extend(ExpressMojitoExtension, libmiddleware, liblogger);
 
 /**
 Creates a new instance of mojito and attach to `app`.
@@ -185,78 +185,6 @@ ExpressMojitoExtension.prototype.createMojito = function (app) {
 };
 
 
-
-/**
-Configures YUI logger to honor the logLevel and logLevelOrder
-TODO: This should be done at the low level in YUI.
-
-@method _configureLogger
-@protected
-@param {YUI} Y shared YUI instance on the server
-**/
-ExpressMojitoExtension.prototype._configureLogger = function (Y) {
-    var logLevel = (Y.config.logLevel || 'debug').toLowerCase(),
-        logLevelOrder = Y.config.logLevelOrder || [],
-        defaultLogLevel = logLevelOrder[0] || 'info',
-        isatty = process.stdout.isTTY;
-
-    function log(c, msg, cat, src) {
-        var f,
-            m = (src) ? src + ': ' + msg : msg;
-
-        // if stdout is bound to the tty, we should try to
-        // use the fancy logs implemented by 'yui-log-nodejs'.
-        // TODO: eventually YUI should take care of this piece.
-        if (isatty && Y.Lang.isFunction(c.logFn)) {
-            c.logFn.call(Y, msg, cat, src);
-        } else if ((typeof console !== 'undefined') && console.log) {
-            f = (cat && console[cat]) ? cat : 'log';
-            console[f](msg);
-        }
-    }
-
-    // one more hack: we need to make sure that base is attached
-    // to be able to listen for Y.on.
-    Y.use('base');
-
-    if (Y.config.debug) {
-
-        logLevel = (logLevelOrder.indexOf(logLevel) >= 0 ?
-                        logLevel : logLevelOrder[0]);
-
-        // logLevel index defines the begining of the logLevelOrder structure
-        // e.g: ['foo', 'bar', 'baz'], and logLevel 'bar' 
-        // should produce: ['bar', 'baz']
-        logLevelOrder = (logLevel ?
-                         logLevelOrder.slice(logLevelOrder.indexOf(logLevel)) :
-                         []);
-
-        Y.applyConfig({
-            useBrowserConsole: false,
-            logLevel: logLevel,
-            logLevelOrder: logLevelOrder
-        });
-
-        // listening for low level log events to filter some of them.
-        Y.on('yui:log', function (e) {
-            var c = Y.config,
-                cat = e && e.cat && e.cat.toLowerCase();
-
-            // this covers the case Y.log(msg) without category
-            // by using the low priority category from logLevelOrder.
-            cat = cat || defaultLogLevel;
-
-            // applying logLevel filters
-            if (cat && ((c.logLevel === cat) ||
-                        (c.logLevelOrder.indexOf(cat) >= 0))) {
-                log(c, e.msg, cat, e.src);
-            }
-            return true;
-        });
-    }
-};
-
-
 /**
  * Configures YUI with both the Mojito framework and all the YUI modules in the
  * application.
@@ -306,13 +234,12 @@ ExpressMojitoExtension.prototype._configureYUI = function (Y, store) {
  */
 ExpressMojitoExtension.prototype._configureAppInstance = function (app, store, options) {
     var my = this,
+        mojitoFn = ExpressMojitoExtension,
         Y,
         appConfig,
         yuiConfig,
         logConfig = {},
         modules = [],
-        middleware,
-        midConfig,
         debugConfig;
 
     if (!options) {
@@ -354,7 +281,7 @@ ExpressMojitoExtension.prototype._configureAppInstance = function (app, store, o
         useSync: true
     });
 
-    my._configureLogger(Y);
+    mojitoFn.configureLogger(Y);
     modules = my._configureYUI(Y, store);
 
     // attaching all modules available for this application for the server side
@@ -362,25 +289,8 @@ ExpressMojitoExtension.prototype._configureAppInstance = function (app, store, o
     Y.use.apply(Y, modules);
     Y.applyConfig({ useSync: false });
 
-    middleware = [].concat(Mojito.MOJITO_MIDDLEWARE);
-
-    midConfig = {
-        Y: Y,
-        store: store,
-        logger: {
-            log: Y.log
-        },
-        context: options.context
-    };
-
-    // expose the middlewares to `app.mojito` 
-    // my._exposeMiddleware(app, _dispatcher(store, appConfig, Y), midConfig, middleware);
-    // ExpressMojitoExtension._exposeMiddleware(app, _dispatcher(store, appConfig, Y), midConfig, middleware);
-    // ExpressMojitoExtension._exposeMiddleware();
-    ExpressMojitoExtension.exposeMiddleware();
-
-    // stash for global access - is that still needed ??
-    // merge(Mojito, midConfig);
+    // expose the middlewares to `mojito` function
+    mojitoFn.exposeMiddleware();
 
     // stash for runtime
     app.set('mojito.store', store);
@@ -416,23 +326,10 @@ appProto.defaultConfiguration = function () {
 //  ----------------------------------------------------------------------------
 
 /**
-For UT, can use the following approach:
 
-    var Mojito = require('mojito'),
-        app = { }; // mocking express
-
-    app.mojito = new Mojito(app);
-    app.mojito._exposeMiddleware = function (app, mid, midConfig, mids) {
-        // assert
-    };
-
-    // create the instance
-    app.mojito.createMojito(app);
-
-    // validate here
+Export Mojito extension as a function.
 
 **/
 
-// exports = module.exports = extend(ExpressMojitoExtension, libmiddleware);
 exports = module.exports = ExpressMojitoExtension;
